+++ client/game/Empire/Planets/Resources/ResourceType.vue
@@ -100,7 +100,7 @@
 
         box-sizing: content-box;
         width: 2.4rem;
-        height: 2.4rem;
+        height: 2.6rem;
         padding: 0.5rem 1rem;
         border: 1px solid palette("grey", "abbey");
         margin: 0 0.4rem 0.4rem 0;